requirements.txt
streamlit>=1.28.0
python-chess>=1.999
numpy>=1.24.0
pandas>=2.0.0
stockfish>=3.28.0
plotly>=5.15.0
streamlit-chess>=1.0.0
